refactor(counter): collapse subtypes into dto classes in entity file
