--- src/inventory/counter/dto/insert-counter-access.dto.ts
@@ -1,8 +1,0 @@
-import { PickType } from '@nestjs/swagger';
-import { CounterAccess } from 'src/inventory/counter/counter-access.entity';
-
-export class InsertCounterAccessDto extends PickType(CounterAccess, [
-  'ruleId',
-  'counterId',
-  'createdBy',
-] as const) {}
