--- src/inventory/counter/dto/insert-counter-entry.dto.ts
@@ -1,8 +1,0 @@
-import { PickType } from '@nestjs/swagger';
-import { CounterEntry } from 'src/inventory/counter/counter-entry.entity';
-
-export class InsertCounterEntryDto extends PickType(CounterEntry, [
-  'counterId',
-  'value',
-  'createdBy',
-] as const) {}
