--- src/inventory/counter/dto/select-counter.dto.ts
@@ -1,4 +1,0 @@
-import { PickType } from '@nestjs/swagger';
-import { Counter } from 'src/inventory/counter/counter.entity';
-
-export class SelectCounterDto extends PickType(Counter, ['id'] as const) {}
